Move code from __init__.py to utils.py
